pyproject.toml
[tool.poetry]
name = "instagram"
version = "0.1.0"
description = ""
authors = ["diko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.3"
djangorestframework = "^3.16.0"
markdown = "^3.8"
django-filter = "^25.1"
dotenv = "^0.9.9"
dj-database-url = "^3.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
